refactor: add baseNotificationService
알림 처리를 위한 기본 클래스 생성
